pyproject.toml
[project]
name = "pricebot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.2",
    "fastapi>=0.122.0",
    "pytelegrambotapi>=4.29.1",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
]
